First version of sourcesmouth-theme after initial clean up.  I know there's more work to be done, but I wanted this published and pushed ASAP.
There is some limited integration with org-publish (for pages).  Also
this release includes Bastien Gurry's blorg.el

http://lumiere.ens.fr/~guerry/blorg.php

This theme is free to use and modify.  This theme is licensed for the
time being under the GPL2/GPL3 license.  The artwork on my blog
http://sourcesmouth.co.uk/blog/ is included but modified with a crude
watermark to remind you to change it.

If this is a problem for you please contact me at
rick@sourcesmouth.co.uk.

Many thanks to the Carsten Dominick, Bastien Gurry, and the org-mode
community who've made this possible.

Additional thanks to David O'Toole for motivating me to publish this.